Comment by ajaxgit add .

+++ database/migrations/2023_04_15_054941_create_comments_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('post_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('user_id');
             $table->text('comment_body');
             $table->timestamps();
         });
